fix(ubsan): silence 5 UBSan-flagged production-code UB sites
1. pdjson.c/h: change stack_top from size_t to ptrdiff_t so that the
   -1 sentinel is a well-defined signed value; update all (size_t)-1
   comparisons to -1 and add casts on the depth/size comparisons to
   keep sign-clean arithmetic.

2. motion_avx512.c (×3 scalar tails): cast uint16_t filter[] operands
   to uint32_t before accumulation; the sum of all five taps at max
   pixel values (≈65536×65535) overflows signed int in the C default
   arithmetic promotions.

3. vif_avx512.c (all 4 sites): cast loop counter i to int before
   subtracting fwidth_half; unsigned − signed promotes to unsigned and
   the subsequent signed-int assignment is implementation-defined when
   the result wraps.

4. adm_avx2.c (10 sites): replace the unsigned hex literal
   0xFFFFFFFFFFFFFFFF passed to _mm256_set_epi64x() with -1LL;
   the hex form overflows long long and is UB per C99 §6.4.4.1.

5. integer_adm.c (both init loops): cast (1u << (shift_flt[idx] - 1))
   to int32_t before assigning to int32_t add_bef_shift_flt[]; the
   1u<<31 wrap is intentional per ADR-0155 (Netflix#955) and is now
   an explicit implementation-defined conversion rather than UB.
   NOLINT annotations cite ADR-0155 inline.

Build: clean (1010/1010 targets). Tests: 87/87 fast suite pass.
No Netflix golden assertion values changed.

+++ core/src/feature/integer_adm.c
@@ -1361,8 +1361,16 @@ static void i4_adm_csf(AdmBuffer *buf, int scale, int w, int h, int stride,
      * or `int64_t` without a coordinated Netflix-side golden-number
      * update. See docs/adr/0155-adm-i4-rounding-deferred-netflix-955.md. */
     for (unsigned idx = 0; idx < 3; ++idx) {
-        add_bef_shift_dst[idx] = (1u << (shift_dst[idx] - 1));
-        add_bef_shift_flt[idx] = (1u << (shift_flt[idx] - 1));
+        /* UBSan: cast unsigned shift result to int32_t explicitly; the wrap
+         * for shift_flt[idx]==32 is intentional per ADR-0155 (Netflix#955). */
+        add_bef_shift_dst[idx] =
+            (int32_t)(1u
+                      << (shift_dst[idx] -
+                          1)); /* NOLINT(hicpp-signed-bitwise) -- ADR-0155: intentional unsigned-to-signed cast to match Netflix golden arithmetic */
+        add_bef_shift_flt[idx] =
+            (int32_t)(1u
+                      << (shift_flt[idx] -
+                          1)); /* NOLINT(hicpp-signed-bitwise) -- ADR-0155: 1u<<31 wraps to INT32_MIN on cast; sign-negated rounding baked into Netflix golden assertions */
     }
 
     /* The computation of the csf values is not required for the regions
@@ -2097,8 +2105,16 @@ static float i4_adm_cm(AdmBuffer *buf, int w, int h, int src_stride, int csf_a_s
      * see note at the earlier `add_bef_shift_flt[]` initialiser loop.
      * Preserved for Netflix-golden bit-exactness. */
     for (unsigned idx = 0; idx < 3; ++idx) {
-        add_bef_shift_dst[idx] = (1u << (shift_dst[idx] - 1));
-        add_bef_shift_flt[idx] = (1u << (shift_flt[idx] - 1));
+        /* UBSan: cast unsigned shift result to int32_t explicitly; the wrap
+         * for shift_flt[idx]==32 is intentional per ADR-0155 (Netflix#955). */
+        add_bef_shift_dst[idx] =
+            (int32_t)(1u
+                      << (shift_dst[idx] -
+                          1)); /* NOLINT(hicpp-signed-bitwise) -- ADR-0155: intentional unsigned-to-signed cast to match Netflix golden arithmetic */
+        add_bef_shift_flt[idx] =
+            (int32_t)(1u
+                      << (shift_flt[idx] -
+                          1)); /* NOLINT(hicpp-signed-bitwise) -- ADR-0155: 1u<<31 wraps to INT32_MIN on cast; sign-negated rounding baked into Netflix golden assertions */
     }
 
     uint32_t shift_cub = (uint32_t)ceil(log2(w));
